Fully implement edge and vertex collisions.
This finishes the physics BSP algorithm (moving bodies still not fully
supported), but there seem to be missing edge and vertex collisions still.
However, its performance is not like the brute force algorithm, which was
unusable for large levels.

A simple test to reproduce the present issue is by resting the ball flat on a
surface, and then moving off an edge.  Debugging showed that the preliminary
edge test flag was set when I did this but no edge collision was produced,
although it's unclear exactly whether these were on the same frame.  Possibly
the early (first) lump guard may be related (or not).

Promising!