Revert "Add a couple tolerances."
Currently, the BSP physics algorithm has a few bugs that need to be fixed.  It
misses some edge collisions if the ball is resting (but can handle it correctly
if the ball falls from a distance onto the edge).  Also, the ball also falls right
through Easy 2 (the tolerances did not fix this issue; I'm reverting them).

This reverts commit a3324293ce98466e943ba215516046244a519fa4.