Provide a fallback implementation for the wires dependency.
This makes the wires dependency optional, but enabled by default.  Because it
is an automatic Cabal flag, Cabal should be able to automatically try enabling
the flag if it can't satisfy the wires dependency.

Currently, the latest wires package on Hackage is not up-to-date with the most
recent versions of its dependencies, so it will fail to build without local
modifications to build a custom package, until a new, up-to-date ‘wires’
package is uploaded to Hackage.